+++ src/org/usfirst/frc/team1351/robot/Drive/DriveTeleOp.java
@@ -11,7 +11,7 @@ public class DriveTeleOp {
 		Drive.setLeft(ControlMode.PercentOutput, value);
 	}
 
-	public static void setGear(byte gear) {
+	public static void setGear(int gear) {
     	Drive.changeGear(gear);
 	}
